Route this branch's three modelling tests to the job that has torch
`test-unit` runs `pytest tests/` minus `.github/ci/unit-test-quarantine.txt`, so a
new test file is gated the day it is committed - and its dependency set has no
torch and no lightgbm. Three of the four files added on this branch reach both:

- `test_latent_convergence_guard` fits a real SDF, CAE and SAE through
  `latent_factors/library_bridge.py` (`import torch` at :11). The determination it
  pins is what those runners record, so a double would not test it.
- `test_fold_seed_coupling` asserts the modules that seed a fold from its index
  reach one derivation, so it imports `tabular_dl` (:41), `deep_learning` (:42) and
  `darts_forecasting` (:20-23), and `gbm` (`import lightgbm` at :40).
- `test_model_catalog_columns` reads `GBM_MODEL_CLASS` and `TABM_MODEL_CLASS` off
  the runners rather than repeating the strings, which is the coupling it exists to
  hold, and both modules are the boundary itself.

The third is why the catalog tests moved out of `tests/test_research_configs.py`
into their own file rather than being deselected by nodeid. That file's other 67
tests need none of this and stay in `test-unit`; quarantining it whole to carry
three would give up their gating, which is the exclude-more-than-you-meant defect
the subtraction exists to prevent. `test_causal_frozen_fraction` stays in
`test-unit` too - `case_studies.utils.causal`, `research.causal` and
`registry.registration` are all already imported by unquarantined causal tests.

Both lists are edited together, which
`test_quarantine_list.py::test_the_image_job_runs_exactly_what_is_quarantined_for_it`
requires: excluded here, run there, or the file runs nowhere. The section header's
file count was stale at 36 against 42 entries before this; it now reads 45.

    uv run pytest tests/test_quarantine_list.py -q
    68 passed

    uv run pytest tests/test_latent_convergence_guard.py \
      tests/test_fold_seed_coupling.py tests/test_model_catalog_columns.py -q
    48 passed

`test-unit`'s own sweep, run the way the job runs it - the quarantine file parsed
into --ignore/--deselect, ML4T_OUTPUT_DIR unset:

    4546 passed, 113 skipped, 101 deselected in 631.38s
