+++ .github/ci/unit-test-quarantine.txt
@@ -17,7 +17,7 @@
 
 # ---------------------------------------------------------------------------
 # Needs the modelling environment - run by test-unit-image, in ml4t/ml4t:latest.
-# 36 files. The 26 that existed on 2026-08-24 measured 392 tests, 3m21s.
+# 45 files. The 26 that existed on 2026-08-24 measured 392 tests, 3m21s.
 #
 # `case_studies/utils/gbm.py` imports torch at module scope, so the LightGBM
 # files are behind this boundary too even though LightGBM itself is a small
@@ -112,6 +112,23 @@ tests/test_tabular_dl_adapter.py
 tests/test_tabular_dl_identity.py
 tests/test_tabular_dl_runtime.py
 tests/test_training_identity_migration.py
+# Three files added on 2026-09-07 with the model-families infrastructure pass, all
+# behind the same import boundary.
+#
+# `test_latent_convergence_guard` fits a real SDF, CAE and SAE through
+# `latent_factors/library_bridge.py`, which imports torch and `ml4t.models` at module
+# scope - the convergence determination it pins is what those runners record, so a double
+# would not test it.
+tests/test_latent_convergence_guard.py
+# `test_fold_seed_coupling` asserts that the four modules seeding a fold from its index
+# reach one derivation, so it imports `tabular_dl`, `deep_learning` and
+# `darts_forecasting`, which reach torch, and `gbm`, which reaches lightgbm.
+tests/test_fold_seed_coupling.py
+# `test_model_catalog_columns` reads `GBM_MODEL_CLASS` and `TABM_MODEL_CLASS` off the
+# runners rather than repeating the strings - which is the coupling it exists to hold -
+# and both modules are the boundary itself. Its sibling `test_research_configs.py` stays
+# in this job: the catalog tests moved out of it for exactly this reason.
+tests/test_model_catalog_columns.py
 
 # ---------------------------------------------------------------------------
 # Executes notebooks through Papermill, against the reader's image and the
